+++ android/src/main/kotlin/com/pravera/flutter_foreground_task/service/ForegroundServiceManager.kt
@@ -18,15 +18,43 @@ class ForegroundServiceManager {
 	 * @param context context
 	 * @param call Method call on the method channel. This includes notification options.
 	 */
-	fun start(context: Context, call: MethodCall) {
-		val intent = Intent(context, ForegroundService::class.java)
-		intent.action = ForegroundServiceAction.START
-		saveOptions(context, call)
+	fun start(context: Context, call: MethodCall): Boolean {
+		try {
+			val intent = Intent(context, ForegroundService::class.java)
+			intent.action = ForegroundServiceAction.START
+			saveOptions(context, call)
 
-		if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O)
-			context.startForegroundService(intent)
-		else
-			context.startService(intent)
+			if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O)
+				context.startForegroundService(intent)
+			else
+				context.startService(intent)
+		} catch (e: Exception) {
+			return false
+		}
+
+		return true
+	}
+
+	/**
+	 * Restart the foreground service.
+	 *
+	 * @param context context
+	 * @param call Method call on the method channel. This includes notification options.
+	 */
+	fun restart(context: Context, call: MethodCall): Boolean {
+		try {
+			val intent = Intent(context, ForegroundService::class.java)
+			intent.action = ForegroundServiceAction.RESTART
+
+			if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O)
+				context.startForegroundService(intent)
+			else
+				context.startService(intent)
+		} catch (e: Exception) {
+			return false
+		}
+
+		return true
 	}
 
 	/**
@@ -35,15 +63,21 @@ class ForegroundServiceManager {
 	 * @param context context
 	 * @param call Method call on the method channel. This includes notification options.
 	 */
-	fun update(context: Context, call: MethodCall) {
-		val intent = Intent(context, ForegroundService::class.java)
-		intent.action = ForegroundServiceAction.UPDATE
-		updateOptions(context, call)
+	fun update(context: Context, call: MethodCall): Boolean {
+		try {
+			val intent = Intent(context, ForegroundService::class.java)
+			intent.action = ForegroundServiceAction.UPDATE
+			updateOptions(context, call)
 
-		if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O)
-			context.startForegroundService(intent)
-		else
-			context.startService(intent)
+			if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O)
+				context.startForegroundService(intent)
+			else
+				context.startService(intent)
+		} catch (e: Exception) {
+			return false
+		}
+
+		return true
 	}
 
 	/**
@@ -51,18 +85,24 @@ class ForegroundServiceManager {
 	 *
 	 * @param context context
 	 */
-	fun stop(context: Context) {
+	fun stop(context: Context): Boolean {
 		// If the service is not running, the stop function is not executed.
-		if (!ForegroundService.isRunningService) return
+		if (!ForegroundService.isRunningService) return false
 
-		val intent = Intent(context, ForegroundService::class.java)
-		intent.action = ForegroundServiceAction.STOP
-		clearOptions(context)
+		try {
+			val intent = Intent(context, ForegroundService::class.java)
+			intent.action = ForegroundServiceAction.STOP
+			clearOptions(context)
 
-		if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O)
-			context.startForegroundService(intent)
-		else
-			context.startService(intent)
+			if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O)
+				context.startForegroundService(intent)
+			else
+				context.startService(intent)
+		} catch (e: Exception) {
+			return false
+		}
+
+		return true
 	}
 
 	/**
@@ -84,6 +124,7 @@ class ForegroundServiceManager {
 		val enableVibration = call.argument<Boolean>(ForegroundServicePrefsKey.ENABLE_VIBRATION) ?: false
 		val playSound = call.argument<Boolean>(ForegroundServicePrefsKey.PLAY_SOUND) ?: false
 		val showWhen = call.argument<Boolean>(ForegroundServicePrefsKey.SHOW_WHEN) ?: false
+		val isSticky = call.argument<Boolean>(ForegroundServicePrefsKey.IS_STICKY) ?: true
 		val visibility = call.argument<Int>(ForegroundServicePrefsKey.VISIBILITY) ?: 1
 
 		val iconData = call.argument<HashMap<String, String>>("iconData")
@@ -106,6 +147,7 @@ class ForegroundServiceManager {
 			putBoolean(ForegroundServicePrefsKey.ENABLE_VIBRATION, enableVibration)
 			putBoolean(ForegroundServicePrefsKey.PLAY_SOUND, playSound)
 			putBoolean(ForegroundServicePrefsKey.SHOW_WHEN, showWhen)
+			putBoolean(ForegroundServicePrefsKey.IS_STICKY, isSticky)
 			putInt(ForegroundServicePrefsKey.VISIBILITY, visibility)
 			putString(ForegroundServicePrefsKey.ICON_RES_TYPE, iconResType)
 			putString(ForegroundServicePrefsKey.ICON_RES_PREFIX, iconResPrefix)