+++ example/lib/main.dart
@@ -104,13 +104,17 @@ class _ExampleAppState extends State<ExampleApp> {
 
   void _startForegroundTask() async {
     // You can save data using the saveData function.
-    await FlutterForegroundTask.saveData('customData', 'hello');
+    await FlutterForegroundTask.saveData(key: 'customData', value: 'hello');
 
-    _receivePort = await FlutterForegroundTask.startService(
-      notificationTitle: 'Foreground Service is running',
-      notificationText: 'Tap to return to the app',
-      callback: startCallback,
-    );
+    if (await FlutterForegroundTask.isRunningService) {
+      _receivePort = await FlutterForegroundTask.restartService();
+    } else {
+      _receivePort = await FlutterForegroundTask.startService(
+        notificationTitle: 'Foreground Service is running',
+        notificationText: 'Tap to return to the app',
+        callback: startCallback,
+      );
+    }
 
     _receivePort?.listen((message) {
       if (message is DateTime)