+++ ios/Classes/service/BackgroundServiceManager.swift
@@ -9,31 +9,51 @@ import Flutter
 import Foundation
 
 class BackgroundServiceManager: NSObject {
-  func start(call: FlutterMethodCall) {
+  func start(call: FlutterMethodCall) -> Bool {
     if #available(iOS 10.0, *) {
       saveOptions(call: call)
       BackgroundService.sharedInstance.run(action: BackgroundServiceAction.START)
     } else {
       // Fallback on earlier versions
+      return false
     }
+    
+    return true
   }
   
-  func update(call: FlutterMethodCall) {
+  func restart(call: FlutterMethodCall) -> Bool {
+    if #available(iOS 10.0, *) {
+      BackgroundService.sharedInstance.run(action: BackgroundServiceAction.RESTART)
+    } else {
+      // Fallback on earlier versions
+      return false
+    }
+    
+    return true
+  }
+  
+  func update(call: FlutterMethodCall) -> Bool {
     if #available(iOS 10.0, *) {
       updateOptions(call: call)
       BackgroundService.sharedInstance.run(action: BackgroundServiceAction.UPDATE)
     } else {
       // Fallback on earlier versions
+      return false
     }
+    
+    return true
   }
   
-  func stop() {
+  func stop() -> Bool {
     if #available(iOS 10.0, *) {
       clearOptions()
       BackgroundService.sharedInstance.run(action: BackgroundServiceAction.STOP)
     } else {
       // Fallback on earlier versions
+      return false
     }
+    
+    return true
   }
   
   func isRunningService() -> Bool {
@@ -54,15 +74,18 @@ class BackgroundServiceManager: NSObject {
     let playSound = argsDict[PLAY_SOUND] as? Bool ?? false
     let taskInterval = argsDict[TASK_INTERVAL] as? Int ?? 5000
     let callbackHandle = argsDict[CALLBACK_HANDLE] as? Int64
-    
+
     prefs?.set(notificationContentTitle, forKey: NOTIFICATION_CONTENT_TITLE)
     prefs?.set(notificationContentText, forKey: NOTIFICATION_CONTENT_TEXT)
     prefs?.set(showNotification, forKey: SHOW_NOTIFICATION)
     prefs?.set(playSound, forKey: PLAY_SOUND)
     prefs?.set(taskInterval, forKey: TASK_INTERVAL)
     prefs?.removeObject(forKey: CALLBACK_HANDLE)
+    prefs?.removeObject(forKey: CALLBACK_HANDLE_ON_RESTART)
+    
     if callbackHandle != nil {
       prefs?.set(callbackHandle, forKey: CALLBACK_HANDLE)
+      prefs?.set(callbackHandle, forKey: CALLBACK_HANDLE_ON_RESTART)
     }
   }
   
@@ -83,6 +106,7 @@ class BackgroundServiceManager: NSObject {
     prefs?.removeObject(forKey: CALLBACK_HANDLE)
     if callbackHandle != nil {
       prefs?.set(callbackHandle, forKey: CALLBACK_HANDLE)
+      prefs?.set(callbackHandle, forKey: CALLBACK_HANDLE_ON_RESTART)
     }
   }