+++ lib/flutter_foreground_task.dart
@@ -56,9 +56,11 @@ class FlutterForegroundTask {
   }) async {
     _androidNotificationOptions = androidNotificationOptions;
     _iosNotificationOptions = iosNotificationOptions ??
-        _iosNotificationOptions ?? const IOSNotificationOptions();
+        _iosNotificationOptions ??
+        const IOSNotificationOptions();
     _foregroundTaskOptions = foregroundTaskOptions ??
-        _foregroundTaskOptions ?? const ForegroundTaskOptions();
+        _foregroundTaskOptions ??
+        const ForegroundTaskOptions();
     _printDevLog = printDevLog ?? _printDevLog;
   }
 
@@ -87,26 +89,50 @@ class FlutterForegroundTask {
     options['notificationContentTitle'] = notificationTitle;
     options['notificationContentText'] = notificationText;
     if (callback != null) {
-      options.addAll(
-          _foregroundTaskOptions?.toJson() ?? Map<String, dynamic>());
+      options.addAll(_foregroundTaskOptions!.toJson());
       options['callbackHandle'] =
           PluginUtilities.getCallbackHandle(callback)?.toRawHandle();
     }
 
-    _methodChannel.invokeMethod('startForegroundService', options);
-    _printMessage('FlutterForegroundTask started.');
+    final bool result =
+        await _methodChannel.invokeMethod('startForegroundService', options);
+    if (result) {
+      _printMessage('FlutterForegroundTask started');
+      return receivePort;
+    }
 
-    return receivePort;
+    return null;
+  }
+
+  /// Restart the foreground service.
+  /// The option value uses the option value of the currently running service as it is.
+  static Future<ReceivePort?> restartService() async {
+    if (!await isRunningService)
+      throw ForegroundTaskException('There are no service started or running.');
+
+    final receivePort = _registerPort();
+    if (receivePort == null)
+      throw ForegroundTaskException(
+          'Failed to register SendPort to communicate with background isolate.');
+
+    final bool result =
+        await _methodChannel.invokeMethod('restartForegroundService');
+    if (result) {
+      _printMessage('FlutterForegroundTask restarted');
+      return receivePort;
+    }
+
+    return null;
   }
 
   /// Update the foreground service.
-  static Future<void> updateService({
+  static Future<bool> updateService({
     String? notificationTitle,
     String? notificationText,
     Function? callback,
   }) async {
     // If the service is not running, the update function is not executed.
-    if (!await isRunningService) return;
+    if (!await isRunningService) return false;
 
     final options = Map<String, dynamic>();
     options['notificationContentTitle'] = notificationTitle;
@@ -116,19 +142,31 @@ class FlutterForegroundTask {
           PluginUtilities.getCallbackHandle(callback)?.toRawHandle();
     }
 
-    _methodChannel.invokeMethod('updateForegroundService', options);
-    _printMessage('FlutterForegroundTask updated.');
+    final bool result =
+        await _methodChannel.invokeMethod('updateForegroundService', options);
+    if (result) {
+      _printMessage('FlutterForegroundTask updated');
+      return true;
+    }
+
+    return false;
   }
 
   /// Stop the foreground service.
-  static Future<void> stopService() async {
+  static Future<bool> stopService() async {
     // If the service is not running, the stop function is not executed.
-    if (!await isRunningService) return;
+    if (!await isRunningService) return false;
 
     _removePort();
 
-    _methodChannel.invokeMethod('stopForegroundService');
-    _printMessage('FlutterForegroundTask stopped.');
+    final bool result =
+        await _methodChannel.invokeMethod('stopForegroundService');
+    if (result) {
+      _printMessage('FlutterForegroundTask stopped');
+      return true;
+    }
+
+    return false;
   }
 
   /// Returns whether the foreground service is running.
@@ -144,7 +182,8 @@ class FlutterForegroundTask {
   }
 
   /// Save data with [key].
-  static Future<bool> saveData(String key, Object value) async {
+  static Future<bool> saveData(
+      {required String key, required Object value}) async {
     final prefs = await SharedPreferences.getInstance();
 
     if (value is int)
@@ -195,14 +234,16 @@ class FlutterForegroundTask {
     // This function only works on Android.
     if (!Platform.isAndroid) return true;
 
-    return await _methodChannel.invokeMethod('openIgnoreBatteryOptimizationSettings');
+    return await _methodChannel
+        .invokeMethod('openIgnoreBatteryOptimizationSettings');
   }
 
   /// Set up the task handler and start the foreground task.
   /// It must always be called from a top-level function, otherwise foreground task will not work.
   static void setTaskHandler(TaskHandler handler) {
     // Create a method channel to communicate with the platform.
-    const _backgroundChannel = MethodChannel('flutter_foreground_task/background');
+    const _backgroundChannel =
+        MethodChannel('flutter_foreground_task/background');
 
     // Binding the framework to the flutter engine.
     WidgetsFlutterBinding.ensureInitialized();