+++ lib/models/android_notification_options.dart
@@ -15,6 +15,7 @@ class AndroidNotificationOptions {
     this.enableVibration = false,
     this.playSound = false,
     this.showWhen = false,
+    this.isSticky = true,
     this.visibility = NotificationVisibility.VISIBILITY_PUBLIC,
     this.iconData,
   });
@@ -51,6 +52,10 @@ class AndroidNotificationOptions {
   /// The default is `false`.
   final bool showWhen;
 
+  /// Whether or not the system will restart the service if the service is killed.
+  /// The default is `true`.
+  final bool isSticky;
+
   /// Control the level of detail displayed in notifications on the lock screen.
   /// The default is `NotificationVisibility.VISIBILITY_PUBLIC`.
   final NotificationVisibility visibility;
@@ -70,6 +75,7 @@ class AndroidNotificationOptions {
       'enableVibration': enableVibration,
       'playSound': playSound,
       'showWhen': showWhen,
+      'isSticky': isSticky,
       'visibility': visibility.rawValue,
       'iconData': iconData?.toJson(),
     };